LayoutTests/fast/scrolling/mac/scrollbars/overflow-overlay-scrollbar-hovered-expected.txt

Test hover over overlay scrollbar

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state
enabled
Hovering vertical scrollbar should show expanded scrollbar
PASS Scrollbar is enabled and expanded
Unhovering vertical scrollbar should hide it
PASS Thumb and track hidden
PASS successfullyParsed is true

TEST COMPLETE
